Introduce Transformer abstraction
I don't think I actually want to know about propertyMappings, it may be
better to have a "createTransformer" DSL that takes in either types or
the concrete instances and removes the boilerplate of CarTransformer and
BananaTransformer.

One possible benefit is to only reveal a TransformersFactory and the
Transformer interface to consumers.